+++ bootify.sh
@@ -1,14 +1,46 @@
 #!/bin/bash
 
-usage() {
-	echo "Usage: $0 -d [device] -s [boot|uefi] -i [ISO]"
+VER="0.2"
+
+usage()
+{
+cat <<EOF
+Usage: $0 -d [DEVICE] -s [boot|uefi] -i [ISO]
+
+Options:
+
+-d [DEVICE]     USB device name, eg /dev/sdb
+-s [boot|uefi]  boot method, 'boot' or 'uefi'
+-i [ISO FILE]   path to the ISO file
+-h              display this help and exit
+-l              display available USB devices and exit
+-v              display version and exit
+EOF
 }
 
-while getopts ":hd:s:i:" OPT
+version()
+{
+	echo "bootify $VER"
+}
+
+available_devices()
+{
+	lsblk -ndo tran,name,vendor,model,size | grep usb | tr -s " "  " "
+}
+
+while getopts ":hlvd:s:i:" OPT
 do
 	case $OPT in
 		h)
 			usage
+			exit 0
+			;;
+		v)
+			version
+			exit 0
+			;;
+		l)
+			available_devices
 			exit 0
 			;;
 		d)
@@ -18,15 +50,15 @@ do
 			SCH="$OPTARG"
 			;;
 		i)
-			ISO=$OPTARG
+			ISO="$OPTARG"
 			;;
 		\?)
-			echo "Invalid parameter: -$OPTARG" >&2
+			echo "Invalid option: -$OPTARG" 1>&2
 			usage
 			exit 1
 			;;
 		:)
-			echo "Parameter -$OPTARG requires an argument" >&2
+			echo "Option -$OPTARG requires an argument" 1>&2
 			usage
 			exit 1
 			;;
@@ -43,51 +75,70 @@ fi
 
 # Check if dependencies are met
 
-type parted > /dev/null 2>&1 || { echo >&2 "parted is not ready"; exit 1; }
-type mkfs.ntfs > /dev/null 2>&1 || { echo >&2 "mkfs.ntfs is not ready"; exit 1; }
-type mkfs.vfat > /dev/null 2>&1 || { echo >&2 "mkfs.vfat is not ready"; exit 1; }
-type lsblk > /dev/null 2>&1 || { echo >&2 "lsblk is not ready"; exit 1; }
-type 7za > /dev/null 2>&1 || { echo >&2 "7za is not ready"; exit 1; }
+DEP="dd isoinfo lsblk mkfs.ntfs mkfs.vfat parted stat 7za"
+for D in $DEP
+do
+    type $D > /dev/null 2>&1 || { echo 1>&2 "$D is not ready"; exit 1; }
+done
 
 # Check if bootsect exists
 
 if [[ ! -f res/bootsect ]]
 then
-	echo "bootsect does not exist"
+	echo "bootsect does not exist" 1>&2
 	exit 1
 fi
 
-# Check if DEVICE is a block device
+# Make sure bootify runs as root
+
+if [[ "$EUID" -ne 0 ]]
+then
+	echo "bootify must be run as root" 1>&2
+	exit 1
+fi
+
+# Check if $DEV is a block device
 
 if [[ ! -b "$DEV" ]]
 then
-	echo "$DEV is not a block device"
+	echo "$DEV is not a block device" 1>&2
 	exit 1
 fi
 
-# Check if DEVICE is mounted
+# Check if $DEV is a usb device
+
+if [[ -z $(lsblk -ndo tran $DEV | grep usb) ]]
+then
+	echo "$DEV is not a USB device" 1>&2
+	exit 1
+fi
+
+# Check if $DEV is mounted
 
 if [[ ! -z $(grep $DEV /proc/mounts) ]]
 then
-	echo "$DEV is mounted, dismount it and run bootify again"
-	exit 0
-fi
-
-# Check if ISO file exists
-# TODO: check if ISO has a UDF filesystem
-
-if [[ ! -f "$ISO" ]]
-then
-	echo "$ISO does not exist"
+	echo "$DEV is mounted, dismount it and run bootify again" 1>&2
 	exit 1
 fi
 
-# Check DEVICE capacity
+# Check if $ISO exists and is valid
 
-if [[ $(du -b $ISO | cut -f1) -gt $(lsblk -ndbo size $DEV) ]]
+if [[ ! -f "$ISO" ]]
 then
-	echo "The device capacity is insufficient"
-	exit 0
+	echo "$ISO does not exist" 1>&2
+	exit 1
+elif [[ -z $(isoinfo -d -i "$ISO" | grep "CD-ROM is in ISO 9660 format") ]]
+then
+	echo "$ISO is not a valid ISO file" 1>&2
+	exit 1
+fi
+
+# Check $DEV capacity
+
+if [[ $(stat -c %s $ISO) -gt $(lsblk -ndbo size $DEV) ]]
+then
+	echo "The device capacity is insufficient" 1>&2
+	exit 1
 fi
 
 # Check if /media directory exists
@@ -101,12 +152,12 @@ mbr_part()
 {
 	parted -s -a optimal $DEV mktable msdos mkpart primary ntfs 1 100% set 1 boot
 	mkfs.ntfs -f -L BOOTIFY ${DEV}1
-	dd if=res/bootsect of=$DEV > /dev/null 2>&1
+	dd if=res/bootsect of=$DEV
 	copy_files
 }
 
 gpt_part()
-{	
+{   
 	parted -s -a optimal $DEV mktable gpt mkpart primary fat32 1 100%
 	mkfs.vfat -F32 -n BOOTIFY ${DEV}1
 	copy_files
@@ -116,20 +167,22 @@ copy_files()
 {
 	mount ${DEV}1 /mnt
 	mount -o loop $ISO /media
-	# TODO: provide a nicer progress indicator
+	# TODO: provide a progress indicator
 	cp -rv /media/* /mnt
 
 	# Windows 7 missing UEFI boot file workaround
-	# This does not happens with Windows 8 installation media
+	# This does not happen with Windows 8 installation media
 
 	if [[ "$SCH" == "uefi" ]]
 	then
 		if [[ ! -d /mnt/efi/boot ]]
 		then
+			PWD=pwd
 			mkdir /mnt/efi/boot
 			cd /mnt/efi/boot
-			7z e /media/sources/install.wim 1/Windows/Boot/EFI/bootmgfw.efi > /dev/null 2>&1
+			7z e /media/sources/install.wim 1/Windows/Boot/EFI/bootmgfw.efi 2>&1
 			mv bootmgfw.efi bootx64.efi
+			cd $PWD
 		fi
 	fi
 
@@ -166,7 +219,7 @@ then
 	confirm
 	gpt_part
 else
-	echo "Incorrect boot method, use 'boot' or 'uefi'"
+	echo "Incorrect boot method, use 'boot' or 'uefi'" 1>&2
 	usage
 	exit 1
 fi