+++ CHANGELOG.md
@@ -1,2 +1,16 @@
-0.1 (June 9, 2015)
-First release
+== HEAD
+
+== 0.2 (June 10,2015)
+ 
+* Add check if running as root
+* Add ISO file verification
+* Add display version option
+* Add list available USB devices option
+* Replace `du` by `stat`
+* Improve dependencies check
+* Improve help
+* Some minor changes
+
+== 0.1 (June 9, 2015)
+
+* First release